requirements.txt
astroid==3.3.9
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dill==0.3.9
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
pyee==12.1.1
pylint==3.3.6
pytest==8.3.5
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.7.0
python-dotenv==1.1.0
python-slugify==8.0.4
requests==2.32.3
text-unidecode==1.3
tomlkit==0.13.2
typing_extensions==4.13.0
urllib3==2.3.0
